+++ VOD/src/main/webapp/WEB-INF/views/memberInfo.jsp
@@ -9,40 +9,7 @@
 
 <!DOCTYPE html>
 <html lang="ko">
-
-<head>
-<meta charset="UTF-8">
-<meta name="description" content="Anime Template">
-<meta name="keywords" content="Anime, unica, creative, html">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta http-equiv="X-UA-Compatible" content="ie=edge">
-<title>NeTupidia</title>
-
-<!-- Google Font -->
-<link
-	href="https://fonts.googleapis.com/css2?family=Oswald:wght@300;400;500;600;700&display=swap"
-	rel="stylesheet">
-<link
-	href="https://fonts.googleapis.com/css2?family=Mulish:wght@300;400;500;600;700;800;900&display=swap"
-	rel="stylesheet">
-
-<!-- Css Styles -->
-<link rel="stylesheet" href="resources/css/bootstrap.min.css"
-	type="text/css">
-<link rel="stylesheet" href="resources/css/font-awesome.min.css"
-	type="text/css">
-<link rel="stylesheet" href="resources/css/elegant-icons.css"
-	type="text/css">
-<link rel="stylesheet" href="resources/css/plyr.css" type="text/css">
-<link rel="stylesheet" href="resources/css/nice-select.css"
-	type="text/css">
-<link rel="stylesheet" href="resources/css/owl.carousel.min.css"
-	type="text/css">
-<link rel="stylesheet" href="resources/css/slicknav.min.css"
-	type="text/css">
-<link rel="stylesheet" href="resources/css/style.css" type="text/css">
-</head>
-
+<%@ include file="./include/head.jsp"%>
 <body>
 	<!-- Page Preloder -->
 	<div id="preloder">
@@ -160,7 +127,9 @@
 	}
 	
 	function fn_Delete(){
-		location.href='${context}/deleteMember.do';
+		if(confirm("삭제를 하면 모든 이력이 사라집니다.\n괜찮으시겠습니까?")){
+			location.href='${context}/deleteMember.do';
+		}
 	}
 	
 	function fn_update(){
